feat(app) - Added url_dialog & an invite option
