+++ lib/widgets/end_drawer/end_drawer.dart
@@ -15,12 +15,14 @@ import 'package:junto_beta_mobile/screens/welcome/bloc/auth_event.dart';
 import 'package:junto_beta_mobile/widgets/avatars/member_avatar.dart';
 import 'package:junto_beta_mobile/widgets/background/background_theme.dart';
 import 'package:junto_beta_mobile/widgets/dialogs/confirm_dialog.dart';
+import 'package:junto_beta_mobile/widgets/dialogs/url_dialog.dart';
 import 'package:junto_beta_mobile/widgets/end_drawer/end_drawer_relationships/end_drawer_relationships.dart';
 import 'package:junto_beta_mobile/widgets/end_drawer/junto_account.dart';
 import 'package:junto_beta_mobile/widgets/fade_route.dart';
 import 'package:junto_beta_mobile/widgets/utils/app_version_label.dart';
 import 'package:provider/provider.dart';
 import 'package:junto_beta_mobile/widgets/drawer/junto_filter_drawer.dart';
+import 'package:url_launcher/url_launcher.dart';
 
 import 'junto_contacts.dart';
 import 'junto_themes_page.dart';
@@ -138,6 +140,29 @@ class _JuntoDrawerState extends State<JuntoDrawer> {
                               CupertinoPageRoute<dynamic>(
                                 builder: (BuildContext context) {
                                   return JuntoContacts();
+                                },
+                              ),
+                            );
+                          },
+                        ),
+                        JuntoDrawerItem(
+                          icon: Icons.mail_outline,
+                          title: 'Feedback',
+                          theme: theme,
+                          onTap: () {
+                            showDialog(
+                              context: context,
+                              builder: (BuildContext context) => UrlDialog(
+                                context: context,
+                                text:
+                                    "Feel free to share your feedback through our typeform!",
+                                urlText: "Share Feedback",
+                                onTap: () async {
+                                  // TODO: need to update the url
+                                  final url = "https://junto.typeform.com/";
+                                  if (await canLaunch(url)) {
+                                    await launch(url);
+                                  }
                                 },
                               ),
                             );
