+++ lib/widgets/end_drawer/junto_invite_appbar.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:junto_beta_mobile/app/custom_icons.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:junto_beta_mobile/models/user_model.dart';
+import 'package:junto_beta_mobile/widgets/dialogs/url_dialog.dart';
 import 'package:share/share.dart';
 
 class JuntoInviteAppBar extends StatelessWidget {
@@ -49,9 +50,17 @@ class JuntoInviteAppBar extends StatelessWidget {
               onTap: () async {
                 showDialog(
                   context: context,
-                  builder: (BuildContext context) => JuntoBetaInviteDialog(
+                  builder: (BuildContext context) => UrlDialog(
                     context: context,
-                    userProfile: userProfile,
+                    text:
+                        "Feel free to use this link to our private beta in case you'd like to email your friends instead!",
+                    urlText: "Copy Link",
+                    onTap: () {
+                      Share.share(
+                        "https://junto.typeform.com/to/k7BUVK8f",
+                        subject: 'Check out the Junto beta!',
+                      );
+                    },
                   ),
                 );
               },
@@ -79,113 +88,6 @@ class JuntoInviteAppBar extends StatelessWidget {
               ),
             ),
           ),
-        ),
-      ),
-    );
-  }
-}
-
-class JuntoBetaInviteDialog extends StatelessWidget {
-  const JuntoBetaInviteDialog({
-    this.context,
-    this.userProfile,
-  });
-
-  final BuildContext context;
-  final UserData userProfile;
-
-  @override
-  Widget build(BuildContext context) {
-    return Dialog(
-      shape: RoundedRectangleBorder(
-        borderRadius: BorderRadius.circular(15),
-      ),
-      backgroundColor: Colors.transparent,
-      child: Container(
-        width: MediaQuery.of(context).size.width,
-        margin: const EdgeInsets.symmetric(horizontal: 10),
-        padding: const EdgeInsets.only(
-          top: 25,
-          left: 25,
-          right: 25,
-        ),
-        decoration: BoxDecoration(
-          color: Theme.of(context).backgroundColor,
-          borderRadius: BorderRadius.circular(15),
-        ),
-        child: Column(
-          mainAxisSize: MainAxisSize.min,
-          children: <Widget>[
-            Text(
-              "Feel free to use this link to our private beta in case you'd like to email your friends instead!'",
-              textAlign: TextAlign.center,
-              overflow: TextOverflow.fade,
-              style: TextStyle(
-                fontSize: 17,
-                color: Theme.of(context).primaryColor,
-                fontWeight: FontWeight.w500,
-              ),
-            ),
-            GestureDetector(
-              onTap: () {
-                Share.share(
-                  "https://junto.typeform.com/to/k7BUVK8f",
-                  subject: 'Check out the Junto beta!',
-                );
-              },
-              child: Container(
-                color: Colors.transparent,
-                padding: const EdgeInsets.symmetric(
-                  vertical: 15,
-                  horizontal: 10,
-                ),
-                child: Row(
-                  mainAxisAlignment: MainAxisAlignment.center,
-                  children: [
-                    Image.asset(
-                      'assets/images/junto-mobile__external-link.png',
-                      height: 17,
-                      color: Theme.of(context).colorScheme.primary,
-                    ),
-                    const SizedBox(width: 5),
-                    Text(
-                      'Copy Link',
-                      style: TextStyle(
-                        fontSize: 17,
-                        fontWeight: FontWeight.w700,
-                        color: Theme.of(context).colorScheme.primary,
-                      ),
-                    ),
-                  ],
-                ),
-              ),
-            ),
-            const SizedBox(height: 15),
-            Row(
-              mainAxisAlignment: MainAxisAlignment.center,
-              children: <Widget>[
-                GestureDetector(
-                  onTap: () {
-                    Navigator.pop(context);
-                  },
-                  child: Container(
-                    padding:
-                        const EdgeInsets.only(right: 25, left: 25, bottom: 25),
-                    color: Colors.transparent,
-                    alignment: Alignment.center,
-                    child: Text(
-                      'CLOSE',
-                      style: TextStyle(
-                        color: Theme.of(context).primaryColor,
-                        fontSize: 12,
-                        fontWeight: FontWeight.w500,
-                      ),
-                    ),
-                  ),
-                ),
-              ],
-            ),
-          ],
         ),
       ),
     );
